Add files via upload
This is the jupyter file for the code for Fine-tuning of Mistral 7B v0.2 by Mistral AI. The model is trained on a dataset contianing historical data for Bills of Quantities codified in accorcance with New Rules of Measurement 1 method of measurement. The goal of this file is to expedite the process of codifying newly received Bills of Quantities and save weeks of working for construction estimators and quantity surveyors.